+++ sources/DirectXFramework/SIG/RT.h
@@ -56,7 +56,7 @@ class RTHolder :public Table
 	{
 		compiled.table_rtv = context.rtv_cpu.place(sizeof(rtv) / sizeof(Render::Handle));
 		auto ptr = reinterpret_cast<Render::Handle*>(&rtv);
-		for (int i = 0; i < compiled.table_rtv.get_count(); i++)
+		for (UINT i = 0; i < (UINT)compiled.table_rtv.get_count(); i++)
 		{
 			Render::Handle* handle = ptr + i;
 			compiled.table_rtv[i].place(*handle, D3D12_DESCRIPTOR_HEAP_TYPE_RTV);
@@ -72,7 +72,7 @@ class RTHolder :public Table
 	{
 		compiled.table_dsv = context.dsv_cpu.place(sizeof(dsv) / sizeof(Render::Handle));
 		auto ptr = reinterpret_cast<Render::Handle*>(&dsv);
-		for (int i = 0; i < compiled.table_dsv.get_count(); i++)
+		for (UINT i = 0; i < (UINT)compiled.table_dsv.get_count(); i++)
 		{
 			Render::Handle* handle = ptr + i;
 			compiled.table_dsv[i].place(*handle, D3D12_DESCRIPTOR_HEAP_TYPE_DSV);
