+++ sources/RenderSystem/Scene/Scene.h
@@ -181,14 +181,14 @@ public:
 
 
 						auto info = frame.place_raw(UINT(data.size()));
-						auto srv = info.resource->create_view<StructuredBufferView<UINT>>(frame, info.offset, info.size).get_srv();
+						auto srv = info.resource->create_view<StructuredBufferView<UINT>>(frame, (UINT)info.offset, (UINT)info.size).get_srv();
 						gather_global.GetMeshes_count() = srv;
                     }
 				
 
                     if (data.size()) {
                         auto info = frame.place_raw(data);
-                        auto srv = info.resource->create_view<FormattedBufferView<UINT, DXGI_FORMAT::DXGI_FORMAT_R32_UINT>>(frame, info.offset, info.size).get_srv();
+                        auto srv = info.resource->create_view<FormattedBufferView<UINT, DXGI_FORMAT::DXGI_FORMAT_R32_UINT>>(frame, (UINT)info.offset, (UINT)info.size).get_srv();
                         gather_global.GetCommands() = srv;
                     }
